examples/part3-multimodal-vision/requirements.txt
openai>=1.0.0
pillow>=10.0.0
streamlit>=1.28.0
